lifted state up to App component from CoinDropDownContainer
