+++ src/containers/coin_drop_down_container.js
@@ -9,31 +9,10 @@ const Wrapper = styled.div`
 `;
 
 class CoinDropDownContainer extends Component {
-    constructor(props) {
-        super(props);
-
-        this.state = { coins: [] };
-    }
-    
-    componentDidMount() {
-        axios.get('https://api.coinmarketcap.com/v1/ticker/')
-            .then((response) => {
-                const coins = response.data.map((coin) => {
-                    return {name: coin.name, id: coin.id};
-                });
-
-                this.setState({coins})
-            })
-            .catch((err) => console.log(err));
-    }
-    
     render() {
-
-        const { coins } = this.state;
-
         return (
             <Wrapper>
-                <CoinDropDown coins={coins}/>
+                <CoinDropDown coins={this.props.coins}/>
             </Wrapper>
         )
     }
